Integrated EventLogger with Level 1
There are a few bugs at the moment.  We should log the instance of
chunks correctly. Right now if the player is hit by a bullet, the memory
is cleared and the instance will increment. We should assign the
instance to the chunk with an unique identifier if the identifier has
not already been assigned an instance number. Else we use the old
instance number. We also cannot log multiple pieces of information on a
single tile e.g. it is impossible to log both drone collision and bullet
collisions, since they are logged at the same position in the memory
grid. One solution could be to have different layers of memory to store
those interactions.